+++ @planetarium/tx/src/tx/index.ts
@@ -4,7 +4,9 @@ export type {
   UnsignedTxWithSystemAction,
   UnsignedTxWithCustomActions,
 } from "./unsigned.js";
+export type { SignedTx } from "./signed.js";
 export {
   encodeUnsignedTxWithSystemAction,
   encodeUnsignedTxWithCustomActions,
 } from "./unsigned.js";
+export { signTx, encodeSignedTx } from "./signed.js";