+++ @planetarium/tx/test/tx/fixtures.ts
@@ -1,3 +1,4 @@
+import { Address, RawPrivateKey } from "@planetarium/account";
 import { fromHex } from "../hex";
 import { Currency } from "../../src/assets";
 
@@ -9,18 +10,28 @@ export const FOO: Currency = {
   maximumSupply: null,
 };
 
+export const account1 = RawPrivateKey.fromHex(
+  "9bf4664ba09a89faeb684b94e69ffde01d26ae14b556204d3f6ab58f61f78418",
+);
+
+export const address1 = Address.deriveFrom(account1);
+
+// FIXME: Replace this with account1/address1:
 export const key1 = {
-  private: fromHex(
-    "9bf4664ba09a89faeb684b94e69ffde01d26ae14b556204d3f6ab58f61f78418"),
-  public: fromHex(
-    "0200e02709cc0c051dc105188c454a2e7ef7b36b85da34529d3abc1968167cf54f"),
-  address: fromHex("268344BA46e6CA2A8a5096565548b9018bc687Ce"),
+  private: account1.toBytes(),
+  public: account1.publicKey.toBytes("compressed"),
+  address: address1.toBytes(),
 };
 
+export const account2 = RawPrivateKey.fromHex(
+  "fcf30b333d04ccfeb562f000a32df488e7154949d31ddcac3cf9278acb5786c7",
+);
+
+export const address2 = Address.deriveFrom(account2);
+
+// FIXME: Replace this with account2/address2:
 export const key2 = {
-  private: fromHex(
-    "fcf30b333d04ccfeb562f000a32df488e7154949d31ddcac3cf9278acb5786c7"),
-  public: fromHex(
-    "0215ba27a461a986f4ce7bcda1fd73dc708da767d0405729edaacaad7b7ff60eed"),
-  address: fromHex("8a29de186B85560D708451101C4Bf02D63b25c50"),
+  private: account2.toBytes(),
+  public: account2.publicKey.toBytes("compressed"),
+  address: address2.toBytes(),
 };