+++ CHANGES.md
@@ -141,6 +141,9 @@ and the specification might change in the near future.
  -  Added `PolymorphicAction<T>.ActionTypeLoader` static property to provide
     a way to configure action type loader to be used in `PolymorphicAction<T>`.
     [[#2873]]
+ -  (@planetarium/tx) Added `SignedTx` type.  [[#2882]]
+ -  (@planetarium/tx) Added `signTx` function.  [[#2882]]
+ -  (@planetarium/tx) Added `encodeSignedTx` function.  [[#2882]]
 
 ### Behavioral changes
 
@@ -167,6 +170,7 @@ and the specification might change in the near future.
 [#2848]: https://github.com/planetarium/libplanet/pull/2848
 [#2872]: https://github.com/planetarium/libplanet/pull/2872
 [#2873]: https://github.com/planetarium/libplanet/pull/2873
+[#2882]: https://github.com/planetarium/libplanet/pull/2882
 [@planetarium/account]: https://www.npmjs.com/package/@planetarium/account